+++ apps/web/src/components/auth/find-email/find-email-form.tsx
@@ -1,19 +1,17 @@
 'use client';
 
 import React, { useState } from 'react';
-import { User, CheckCircle } from 'lucide-react';
 
-import InputIcon from '@/components/common/input/icon';
-import SubmitButton from '@/components/auth/submit-button';
+import { User } from 'lucide-react';
+
 import ErrorMessage from '@/components/auth/error-message';
+import FindEmailSuccess from '@/components/auth/find-email/find-email-success';
+import SubmitButton from '@/components/auth/submit-button';
+import InputIcon from '@/components/common/input/icon';
 
-type Step = 1 | 3;
+type Step = 1 | 2;
 
-interface FindEmailFormProps {
-  onSuccess?: (userInfo: { username: string; email: string }) => void;
-}
-
-const FindEmailForm = ({ onSuccess }: FindEmailFormProps) => {
+const FindEmailForm = () => {
   // 단계 관리
   const [currentStep, setCurrentStep] = useState<Step>(1);
 
@@ -54,10 +52,6 @@ const FindEmailForm = ({ onSuccess }: FindEmailFormProps) => {
     setFieldErrors({});
 
     try {
-      // TODO: 실제 수파베이스 연동 시 여기에 구현
-      // const result = await findEmailByName(formData.name);
-
-      // 임시 더미 로직 (50% 확률로 성공/실패)
       await new Promise((resolve) => setTimeout(resolve, 1500)); // 로딩 시뮬레이션
 
       const isSuccess = Math.random() > 0.5; // 임시 성공/실패 로직
@@ -70,8 +64,7 @@ const FindEmailForm = ({ onSuccess }: FindEmailFormProps) => {
         };
 
         setUserInfo(result);
-        setCurrentStep(3);
-        onSuccess?.(result);
+        setCurrentStep(2);
       } else {
         // 실패 시 에러 메시지 설정
         setFieldErrors({ name: '유저 정보가 없습니다.' });
@@ -84,20 +77,13 @@ const FindEmailForm = ({ onSuccess }: FindEmailFormProps) => {
     }
   };
 
-  // 다시 시도
-  const handleRetry = () => {
-    setCurrentStep(1);
-    setFormData({ name: '' });
-    setFieldErrors({});
-  };
-
   // 폼 유효성 검사
   const isFormValid = (): boolean => {
     return !!formData.name.trim();
   };
 
-  // Step 1 렌더링 (입력 단계)
-  const renderStep1 = () => (
+  //입력 단계
+  const renderInputForm = () => (
     <form onSubmit={handleFormSubmit} className="flex flex-col items-center">
       {/* 이름 입력 필드 */}
       <div
@@ -133,45 +119,12 @@ const FindEmailForm = ({ onSuccess }: FindEmailFormProps) => {
     </form>
   );
 
-  // Step 3 렌더링 (성공)
-  const renderStep3 = () => (
-    <div className="flex flex-col items-center">
-      {/* 성공 아이콘 */}
-      <div className="bg-success-50 mb-6 flex h-16 w-16 items-center justify-center rounded-full">
-        <CheckCircle className="text-success-500 h-8 w-8" />
-      </div>
-
-      {/* 결과 메시지 */}
-      <div className="mb-6 text-center">
-        <p className="text-base text-neutral-600">
-          <span className="font-medium text-neutral-900">{userInfo.username}</span>님이 가입하신
-          이메일은
-        </p>
-        <p className="text-primary-500 mt-1 text-lg font-semibold">{userInfo.email}</p>
-        <p className="mt-1 text-base text-neutral-600">입니다.</p>
-      </div>
-
-      {/* 안내 메시지 */}
-      <div className="flex items-center justify-center">
-        <div className="flex h-4 w-4 items-center justify-center rounded-full bg-neutral-200">
-          <span className="text-xs text-neutral-600">!</span>
-        </div>
-        <span className="ml-2 text-sm text-neutral-600">위의 이메일로 로그인하실 수 있습니다.</span>
-      </div>
-
-      {/* 다시 찾기 버튼 */}
-      <button
-        type="button"
-        onClick={handleRetry}
-        className="hover:text-primary-500 mt-6 text-sm text-neutral-600 transition-colors"
-      >
-        다른 정보로 다시 찾기
-      </button>
-    </div>
-  );
-
   // 현재 단계에 따른 렌더링
-  return currentStep === 1 ? renderStep1() : renderStep3();
+  if (currentStep === 2) {
+    return <FindEmailSuccess userInfo={userInfo} />;
+  }
+
+  return renderInputForm();
 };
 
 export default FindEmailForm;
